Move MEASURE_TAGS to own file
